detector/requirements.txt
fastapi
uvicorn
pyyaml
psutil
watchdog

